+++ include/defs.h
@@ -22,9 +22,9 @@
 #define PAGE_SHIFT      12
 #define PAGE_SIZE       (1 << PAGE_SHIFT)
 #define PAGE_MASK       (~(PAGE_SIZE - 1))
-#define PAGE_UP(p)      (((u64)(p) + PAGE_SIZE - 1) & PAGE_MASK)
-#define PAGE_DOWN(p)    ((u64)(p) & ~(PAGE_SIZE - 1))
-#define PAGE_PFN(p)     ((u64)(p) >> PAGE_SHIFT)
+#define PAGE_UP(p)      (((u32)(p) + PAGE_SIZE - 1) & PAGE_MASK)
+#define PAGE_DOWN(p)    ((u32)(p) & ~(PAGE_SIZE - 1))
+#define PAGE_PFN(p)     ((u32)(p) >> PAGE_SHIFT)
 
 #define GIGABYTE    0x40000000
